refactor: move server tests alongside source files
